buffer: change size field type to size_t

+++ libnopegl/src/backends/gl/buffer_gl.h
@@ -35,9 +35,9 @@ struct buffer_gl {
 struct gpu_ctx;
 
 struct buffer *ngli_buffer_gl_create(struct gpu_ctx *gpu_ctx);
-int ngli_buffer_gl_init(struct buffer *s, int size, int usage);
-int ngli_buffer_gl_upload(struct buffer *s, const void *data, int size, int offset);
-int ngli_buffer_gl_map(struct buffer *s, int size, int offset, void **datap);
+int ngli_buffer_gl_init(struct buffer *s, size_t size, int usage);
+int ngli_buffer_gl_upload(struct buffer *s, const void *data, size_t size, size_t offset);
+int ngli_buffer_gl_map(struct buffer *s, size_t size, size_t offset, void **datap);
 void ngli_buffer_gl_unmap(struct buffer *s);
 void ngli_buffer_gl_freep(struct buffer **sp);
