+++ libnopegl/src/buffer.c
@@ -29,17 +29,17 @@ struct buffer *ngli_buffer_create(struct gpu_ctx *gpu_ctx)
     return gpu_ctx->cls->buffer_create(gpu_ctx);
 }
 
-int ngli_buffer_init(struct buffer *s, int size, int usage)
+int ngli_buffer_init(struct buffer *s, size_t size, int usage)
 {
     return s->gpu_ctx->cls->buffer_init(s, size, usage);
 }
 
-int ngli_buffer_upload(struct buffer *s, const void *data, int size, int offset)
+int ngli_buffer_upload(struct buffer *s, const void *data, size_t size, size_t offset)
 {
     return s->gpu_ctx->cls->buffer_upload(s, data, size, offset);
 }
 
-int ngli_buffer_map(struct buffer *s, int size, int offset, void **datap)
+int ngli_buffer_map(struct buffer *s, size_t size, size_t offset, void **datap)
 {
     return s->gpu_ctx->cls->buffer_map(s, size, offset, datap);
 }
